file seperation and idpwd generation
